PR feedback: extract utils, remove unnecessary state, some code cleanup
